+++ R/mdirt.R
@@ -334,6 +334,7 @@ mdirt <- function(data, model, customTheta = NULL, structure = NULL, item.Q = NU
     if(is.null(group)) group <- rep('all', nrow(data))
     if((is(model, 'mirt.model') || is.character(model)) && is.null(technical$customTheta))
         stop('customTheta input required when using a mirt.model type input')
+    technical$omp <- FALSE
     mods <- myLapply(1:nruns, function(x, ...) return(ESTIMATION(...)), method=method,
                      latent.regression=latent.regression, structure=structure,
                      data=data, model=model, group=group, itemtype=itemtype, optimizer=optimizer,