+++ R/mirtCluster.R
@@ -10,6 +10,8 @@
 #' @aliases mirtCluster
 #' @param spec input that is passed to \code{parallel::makeCluster()}. If no input is given the
 #'   maximum number of available local cores will be used
+#' @param omp_threads number of OpenMP threads to use (currently applies to E-step computations only).
+#'   Not used when argument input is missing
 #' @param ... additional arguments to pass to \code{parallel::makeCluster}
 #' @param remove logical; remove previously defined \code{mirtCluster()}?
 #'
@@ -27,14 +29,19 @@
 #' #make 4 cores available for parallel computing
 #' mirtCluster(4)
 #'
-#' #' #stop and remove cores
+#' #stop and remove cores
 #' mirtCluster(remove = TRUE)
 #'
+#' # create 3 core achitecture in R, and 4 thread architecture with OpenMP
+#' mirtCluster(spec = 3, omp_threads = 4)
+#'
 #' }
-mirtCluster <- function(spec, ..., remove = FALSE){
+mirtCluster <- function(spec, omp_threads, remove = FALSE, ...){
     if(requireNamespace("parallel", quietly = TRUE)){
         if(missing(spec))
             spec <- parallel::detectCores()
+        if(missing(omp_threads))
+            .mirtClusterEnv$omp_threads <- 1L
         if(remove){
             if(is.null(.mirtClusterEnv$MIRTCLUSTER)){
                 message('There is no visible mirtCluster() definition')
@@ -43,6 +50,7 @@ mirtCluster <- function(spec, ..., remove = FALSE){
             parallel::stopCluster(.mirtClusterEnv$MIRTCLUSTER)
             .mirtClusterEnv$MIRTCLUSTER <- NULL
             .mirtClusterEnv$ncores <- 1L
+            .mirtClusterEnv$omp_threads <- 1L
             return(invisible())
         }
         if(!is.null(.mirtClusterEnv$MIRTCLUSTER)){