+++ NEWS.md
@@ -1,5 +1,9 @@
 # Changes in mirt 1.32
 
+- OpenMP support added to E-step portion of the package, where number of threads can be 
+  specified via the `mirtCluster()` function argument `omp_threads`. Special thanks to 
+  Matthias von Davier for providing the `omp reduction` code in the `Estep.cpp` file
+
 - Behaviour of `mirt(..., large)` has now been modified, where `large = TRUE` now skips computing
   the unique response patterns for datasets that likely contain little to no repeated response patterns
   (suggested by Matthias von Davier). The previous two-step behaviour is now acheive by passing